Move graph code to new file
