+++ src/lib/CMakeLists.txt
@@ -300,6 +300,8 @@ set(
     optimizer/strategy/column_pruning_rule.hpp
     optimizer/strategy/dependent_group_by_reduction_rule.cpp
     optimizer/strategy/dependent_group_by_reduction_rule.hpp
+    optimizer/strategy/dips_pruning_graph.cpp
+    optimizer/strategy/dips_pruning_graph.hpp
     optimizer/strategy/dips_pruning_rule.cpp
     optimizer/strategy/dips_pruning_rule.hpp
     optimizer/strategy/expression_reduction_rule.cpp
